+++ mmhuman3d/data/data_converters/expose_curated_fits.py
@@ -12,7 +12,7 @@ from mmhuman3d.core.conventions.keypoints_mapping import (
     convert_kps,
     get_keypoint_idxs_by_part
 )
-from mmhuman3d.data.datasets.pipelines.expose_transforms import keyps_to_bbox
+
 @DATA_CONVERTERS.register_module()
 class ExposeCuratedFitsConverter(BaseModeConverter):
     """Curated fits dataset for ExPose
@@ -97,7 +97,7 @@ class ExposeCuratedFitsConverter(BaseModeConverter):
             kps[RIGHT_HAND_IDXS, -1] = right_hand_conf
             kps[FACE_IDXS, -1] = face_conf
             conf = kps[:,-1]
-            bbox = keyps_to_bbox(kps[:,:2],conf, scale=1.2)
+            bbox = self._keypoints_to_scaled_bbox(kps[:,:2][conf>0], scale=1.2)
             bbox_xywh = self._xyxy2xywh(bbox)
             bbox_xywh_.append(bbox_xywh)
