+++ mmhuman3d/data/datasets/pipelines/__init__.py
@@ -27,8 +27,6 @@ from .transforms import (
     Normalize,
     RandomChannelNoise,
     RandomHorizontalFlip,
-)
-from .expose_transforms import (
     BBoxCenterJitter,
     SimulateLowRes,
 )
