+++ mmhuman3d/models/utils/__init__.py
@@ -1,10 +1,10 @@
 from .fits_dict import FitsDict
 from .inverse_kinematics import batch_inverse_kinematics_transform
 from .res_layer import ResLayer, SimplifiedBasicBlock
-from .SMPLX import SmplxHandCropFunc, SmplxFaceCropFunc,SmplxFaceMergeFunc,SmplxHandMergeFunc
+from .SMPLX import SMPLXHandCropFunc, SMPLXFaceMergeFunc,SMPLXFaceCropFunc,SMPLXHandMergeFunc
 
 __all__ = [
     'FitsDict', 'ResLayer', 'SimplifiedBasicBlock',
     'batch_inverse_kinematics_transform',
-    'SmplxHandCropFunc','SmplxFaceCropFunc','SmplxFaceMergeFunc','SmplxHandMergeFunc'
+    'SMPLXHandCropFunc','SMPLXFaceMergeFunc','SMPLXFaceCropFunc','SMPLXHandMergeFunc'
 ]
